requirements.txt
django==1.10
djangorestframework
django-filter
django-webpack-loader
